+++ lib/src/main/java/com/auth0/jwt/android/impl/HeaderSerializer.java
@@ -1,0 +1,10 @@
+package com.auth0.jwt.android.impl;
+
+/**
+ * Responsible for serializing a JWT's header representation to JSON.
+ */
+public class HeaderSerializer extends ClaimsSerializer<HeaderClaimsHolder> {
+    public HeaderSerializer() {
+        super(HeaderClaimsHolder.class);
+    }
+}